requirements4.txt
selenium
pandas 
mysql-connector-python
fastapi
uvicorn